requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cssutils==2.9.0
cycler==0.12.1
dataframe-image==0.2.3
defusedxml==0.7.1
discord==2.3.2
discord.py==2.3.2
fastjsonschema==2.19.1
fonttools==4.47.0
frozenlist==1.4.1
html2image==2.0.4.3
idna==3.6
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mistune==3.0.2
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
Pillow==10.1.0
platformdirs==4.1.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
soupsieve==2.5
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
websocket-client==1.7.0
yarl==1.9.4